+++ src/components/input/input-base.ts
@@ -8,7 +8,6 @@ import { blazorSuppress } from '../common/decorators/blazorSuppress.js';
 import { Constructor } from '../common/mixins/constructor.js';
 import { EventEmitterMixin } from '../common/mixins/event-emitter.js';
 import { SizableMixin } from '../common/mixins/sizable.js';
-import { Direction } from '../common/types.js';
 import { createCounter, partNameMap } from '../common/util.js';
 import { styles } from './themes/light/input.base.css.js';
 import { styles as bootstrap } from './themes/light/input.bootstrap.css.js';
@@ -62,14 +61,6 @@ export abstract class IgcInputBaseComponent extends FormAssociatedRequiredMixin(
   protected helperText!: Array<HTMLElement>;
 
   /**
-   * The direction attribute of the control.
-   * @attr
-   */
-  @property({ reflect: true })
-  @blazorSuppress()
-  public override dir: Direction = 'auto';
-
-  /**
    * Whether the control will have outlined appearance.
    * @attr
    */
@@ -80,8 +71,24 @@ export abstract class IgcInputBaseComponent extends FormAssociatedRequiredMixin(
    * Makes the control a readonly field.
    * @attr
    */
-  @property({ reflect: true, type: Boolean })
-  public readonly = false;
+  @property({ type: Boolean, reflect: true })
+  public readOnly = false;
+
+  /**
+   * Makes the control a readonly field.
+   * @attr
+   *
+   * @deprecated - since v4.4.0
+   * Use the `readOnly` property instead.
+   */
+  @property({ attribute: false })
+  public get readonly() {
+    return this.readOnly;
+  }
+
+  public set readonly(value: boolean) {
+    this.readOnly = value;
+  }
 
   /**
    * The placeholder attribute of the control.